requirements.txt
wget==3.2
requests==2.28.2
openai==0.26.5
Pillow==9.4.0
opencv-python==4.7.0.72